backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# HTTP and API clients
requests==2.31.0
httpx==0.25.2

# Environment and configuration
python-dotenv==1.0.0

# AI and ML dependencies
langchain==0.1.0
langchain-community==0.0.20
pinecone-client==2.2.4
openai==1.3.0
reportlab==4.0.4

# Data processing
numpy==1.24.3
pandas==2.0.3

# Graph visualization
networkx==3.1
matplotlib==3.6.3
plotly==5.15.0
kaleido==0.2.1
pillow==9.5.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
